Split test_pyinaturalist into one test module per source module
